Rename main folder and split main logic to class and script
