+++ create_redist.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 OUT_FILE=${SCRIPT_DIR}/rtilogparser
-pushd "$SCRIPT_DIR/src"
+pushd "$SCRIPT_DIR"
 
 # Create a zip file with all the python files.
 # The python interpreter is able to read zip files and it will execute the
 # content from __main__.py
 zip -r "${OUT_FILE}.zip" . -i '*.py'
-cd .. && zip "${OUT_FILE}.zip" LICENSE
+zip "${OUT_FILE}.zip" LICENSE
 
 echo '#!/usr/bin/env python' | cat - "${OUT_FILE}.zip" > "${OUT_FILE}"
 chmod +x "${OUT_FILE}"
