+++ logparser/custom/custom.py
@@ -1,4 +1,3 @@
-#!/bin/python
 # Log Parser for RTI Connext.
 #
 #   Copyright 2016 Real-Time Innovations, Inc.
@@ -14,8 +13,15 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-"""Application entry point."""
-from rtilogparser import main
+"""Log parsing functions for custom logs.
 
-if __name__ == "__main__":
-    main()
+Functions:
+  + on_custom_log: Parse a log with a custom prefix.
+"""
+from __future__ import absolute_import
+from logparser.devices.logger import log_event
+
+
+def on_custom_log(match, state):
+    """Parse a log with a custom prefix."""
+    log_event("[App] " + match[0], state)
