+++ logparser/devices/formatdevice.py
