+++ logparser/devices/inputdevices.py
@@ -22,11 +22,11 @@ Classes:
   + InputConsoleDevice: Reads the DDS log messages from the standard input.
   + InputFileDevice: Reads the DDS log messages from a file.
 """
-from __future__ import print_function
+from __future__ import absolute_import, print_function
 from os import fstat
 from sys import stdin, stdout
 from time import time
-from devices.outputdevices import OutputConsoleDevice
+from logparser.devices.outputdevices import OutputConsoleDevice
 
 
 class InputDevice(object):
