+++ logparser/devices/logger.py
