+++ logparser/devices/outputdevices.py
