+++ logparser/events/logs.py
@@ -19,7 +19,7 @@ Functions:
   + get_regex_list: Get the regular expressions and function list.
 """
 from __future__ import absolute_import
-import events.events as events
+import logparser.events.events as events
 
 
 def get_regex_list():
